# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgentHub AI - Multi-tenant SaaS API for AI-powered messaging agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/brains:
    get:
      operationId: listBrains
      tags: [brains]
      summary: List AI brains for a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of brains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiBrain"
    post:
      operationId: createBrain
      tags: [brains]
      summary: Create a new AI brain
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiBrainInput"
      responses:
        "201":
          description: Created brain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiBrain"
  /brains/{brainId}:
    get:
      operationId: getBrain
      tags: [brains]
      summary: Get an AI brain
      parameters:
        - name: brainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiBrain"
        "404":
          description: Not found
    patch:
      operationId: updateBrain
      tags: [brains]
      summary: Update an AI brain
      parameters:
        - name: brainId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiBrainUpdate"
      responses:
        "200":
          description: Updated brain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiBrain"
        "404":
          description: Not found
    delete:
      operationId: deleteBrain
      tags: [brains]
      summary: Delete an AI brain
      parameters:
        - name: brainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Knowledge Items ──────────────────────────────────────────────────────
  /brains/{brainId}/knowledge:
    get:
      operationId: listKnowledgeItems
      tags: [brains]
      summary: List knowledge items for a brain
      parameters:
        - name: brainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of knowledge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
    post:
      operationId: createKnowledgeItem
      tags: [brains]
      summary: Add a knowledge item to a brain
      parameters:
        - name: brainId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeItemInput"
      responses:
        "201":
          description: Created knowledge item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{itemId}:
    delete:
      operationId: deleteKnowledgeItem
      tags: [brains]
      summary: Delete a knowledge item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Agent Chat (test) ────────────────────────────────────────────────────
...
components:
  schemas:
...
    AiBrain:
      type: object
      required: [id, workspaceId, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        agentId:
          type: ["integer", "null"]
        name:
          type: string
        systemPrompt:
          type: ["string", "null"]
        fallbackMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AiBrainInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        agentId:
          type: integer
        systemPrompt:
          type: string
        fallbackMessage:
          type: string
    AiBrainUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        agentId:
          type: integer
        systemPrompt:
          type: string
        fallbackMessage:
          type: string

    # ─── Knowledge Items ──────────────────────────────────────────────────────
    KnowledgeItem:
      type: object
      required: [id, workspaceId, type, title, content, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        brainId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [text, url, faq]
        title:
          type: string
        content:
          type: string
        sourceUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    KnowledgeItemInput:
      type: object
      required: [title, content]
      properties:
        type:
          type: string
          enum: [text, url, faq]
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        sourceUrl:
          type: string

    # ─── Chat ─────────────────────────────────────────────────────────────────
...