# lib/api-zod/src/generated/types/aiBrain.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AgentHub AI - Multi-tenant SaaS API for AI-powered messaging agents
 * OpenAPI spec version: 0.1.0
 */

export interface AiBrain {
  id: number;
  workspaceId: number;
  /** @nullable */
  agentId?: number | null;
  name: string;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  fallbackMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}