# lib/api-zod/src/generated/types/aiBrainInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AgentHub AI - Multi-tenant SaaS API for AI-powered messaging agents
 * OpenAPI spec version: 0.1.0
 */

export interface AiBrainInput {
  /** @minLength 1 */
  name: string;
  agentId?: number;
  systemPrompt?: string;
  fallbackMessage?: string;
}