# lib/api-zod/src/generated/types/knowledgeItemInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AgentHub AI - Multi-tenant SaaS API for AI-powered messaging agents
 * OpenAPI spec version: 0.1.0
 */
import type { KnowledgeItemInputType } from './knowledgeItemInputType';

export interface KnowledgeItemInput {
  type?: KnowledgeItemInputType;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  sourceUrl?: string;
}